+++ spacy/tests/lang/es/test_text.py
@@ -1,5 +1,16 @@
 import pytest
 from spacy.lang.es.lex_attrs import like_num
+from spacy.lang.es import Spanish
+
+
+@pytest.mark.issue(3803)
+def test_issue3803():
+    """Test that spanish num-like tokens have True for like_num attribute."""
+    nlp = Spanish()
+    text = "2 dos 1000 mil 12 doce"
+    doc = nlp(text)
+
+    assert [t.like_num for t in doc] == [True, True, True, True, True, True]
 
 
 def test_es_tokenizer_handles_long_text(es_tokenizer):
