+++ spacy/tests/lang/sv/test_exceptions.py
@@ -1,6 +1,5 @@
 import pytest
 
-
 SV_TOKEN_EXCEPTION_TESTS = [
     (
         "Smörsåsen används bl.a. till fisk",
@@ -15,6 +14,26 @@ SV_TOKEN_EXCEPTION_TESTS = [
         ["Anders", "I.", "tycker", "om", "ord", "med", "i", "i", "."],
     ),
 ]
+
+
+@pytest.mark.issue(805)
+@pytest.mark.parametrize(
+    "text,expected_tokens",
+    [
+        (
+            "Smörsåsen används bl.a. till fisk",
+            ["Smörsåsen", "används", "bl.a.", "till", "fisk"],
+        ),
+        (
+            "Jag kommer först kl. 13 p.g.a. diverse förseningar",
+            ["Jag", "kommer", "först", "kl.", "13", "p.g.a.", "diverse", "förseningar"],
+        ),
+    ],
+)
+def test_issue805(sv_tokenizer, text, expected_tokens):
+    tokens = sv_tokenizer(text)
+    token_list = [token.text for token in tokens if not token.is_space]
+    assert expected_tokens == token_list
 
 
 @pytest.mark.parametrize("text,expected_tokens", SV_TOKEN_EXCEPTION_TESTS)
